+++ packages/use-container-query/rollup/rollup.shared.js
@@ -1,0 +1,8 @@
+import babel from "rollup-plugin-babel";
+import { terser } from "rollup-plugin-terser";
+
+export default {
+  input: __dirname + "/../src/index.js",
+  plugins: [babel(), terser()],
+  external: ["react", "react-dom", "@zeecoder/container-query"]
+};